+++ src/components/pages/HouseDetail.tsx
@@ -11,7 +11,7 @@ import { UserAtom } from '@/stores/auth.store';
 import CommentTemplate from '@/components/templates/CommentTemplate';
 import { houseCommentQuery } from '@/hooks/useCommentReply';
 import { CommentType } from '@/types/houseComment.type';
-import Loading from '@/components/pages/Loading';
+import Loading from '@/components/pages/maintenance/Loading';
 import { routePaths } from '@/constants/route';
 import Container from '@/components/atoms/Container';
 import WithSuspenseAndErrorBoundary from '@/components/molecules/WithSuspenseAndErrorBoundary';
